rename context to usercontext
